+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: letsaguiar <letsaguiar@yandex.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 08:38:45 by letsaguiar        #+#    #+#             */
-/*   Updated: 2024/04/16 08:39:35 by letsaguiar       ###   ########.fr       */
+/*   Updated: 2024/04/17 10:17:01 by letsaguiar       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int    ft_strlen(const char *str)
+#include "libft.h"
+
+size_t  ft_strlen(const char *str)
 {
-    unsigned int    i;
+    size_t  i;
 
     i = 0;
     while (str[i])
