+++ ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: letsaguiar <letsaguiar@yandex.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 11:09:23 by letsaguiar        #+#    #+#             */
-/*   Updated: 2024/04/16 11:38:15 by letsaguiar       ###   ########.fr       */
+/*   Updated: 2024/04/17 10:19:22 by letsaguiar       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
     return(ft_memcmp(s1, s2, n));
 }
